backend/requirements.txt
fastapi
python-socketio[asyncio]
sqlalchemy
openai
open-interpreter
uvicorn
python-socketio
pydantic